+++ client/src/features/facets/Checkbox.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useNavigate, useLocation, useParams } from 'react-router-dom'
 
-import { facetLabels, facetSearchTerms } from '../../config/facets'
+import { facetSearchTerms } from '../../config/facets'
 import { buildQuery, getFacetLabel } from '../../lib/facets/helper'
 import { removeFacet } from '../../lib/facets/removeFacet'
 import { stripYaleIdPrefix } from '../../lib/parse/data/helper'
@@ -79,7 +79,7 @@ const Checkbox: React.FC<IProps> = ({
     dispatch(
       addFacets({ facetName: facetSection, facetUri: event.target.value }),
     )
-    pushSiteImproveEvent('Facets', `Remove ${label}`, facetLabels[facetSection])
+    pushSiteImproveEvent('Facets Checkbox', 'Unchecked', `Facet ${label}`)
     navigate(`${pathname}?${newSearchParams}`)
   }
 
@@ -96,7 +96,7 @@ const Checkbox: React.FC<IProps> = ({
     params.set('facetRequest', 'true')
 
     dispatch(addFacets({ facetName: facetSection, facetUri: strValue }))
-    pushSiteImproveEvent('Facets', `Select ${label}`, facetLabels[facetSection])
+    pushSiteImproveEvent('Facets Checkbox', 'Checked', `Facet ${label}`)
     navigate(`${pathname}?${params.toString()}`)
   }