+++ client/src/features/relatedLists/ListItem.tsx
@@ -21,6 +21,7 @@ interface IProps {
   searchTerm: string | Array<string>
   tab: string
   index: number
+  title: string
   itemSpacing?: string
 }
 
@@ -41,6 +42,7 @@ const ListItem: React.FC<IProps> = ({
   searchTerm,
   tab,
   index,
+  title,
   itemSpacing = 'single',
 }) => {
   const [recordLinkHas404, setRecordLinkHas404] = useState<boolean>(false)
@@ -63,7 +65,11 @@ const ListItem: React.FC<IProps> = ({
     >
       <Col md={9} lg={8} xl={9}>
         <DescriptionTerm>
-          <RecordLink url={uri} returns404={setRecordLinkHas404} />
+          <RecordLink
+            url={uri}
+            returns404={setRecordLinkHas404}
+            linkCategory="Accordion"
+          />
         </DescriptionTerm>
       </Col>
       <Col md={3} lg={4} xl={3}>
@@ -79,9 +85,9 @@ const ListItem: React.FC<IProps> = ({
             }}
             onClick={() =>
               pushSiteImproveEvent(
-                `${uri} Faceted Results Search Link`,
+                'Search Link',
                 'Selected',
-                linkLabel,
+                `Accordion ${title}`,
               )
             }
             data-testid={`list-item-link-${index}`}