+++ client/src/features/relatedLists/SearchResultRelatedList.tsx
@@ -15,6 +15,7 @@ import { pushSiteImproveEvent } from '../../lib/siteImprove'
 interface IProps {
   url: string
   data: ISearchResults
+  title: string
   scope?: string
 }
 
@@ -34,13 +35,22 @@ const QueryRelationsListRow: React.FC<{ uri: string; index: number }> = ({
       data-testid={`query-relations-list-row-${index}`}
     >
       <Col xs={12} className="justify-content-start">
-        <RecordLink url={uri} returns404={setRecordLinkHas404} />
+        <RecordLink
+          url={uri}
+          returns404={setRecordLinkHas404}
+          linkCategory="Accordion"
+        />
       </Col>
     </Row>
   )
 }
 
-const SearchResultRelatedList: React.FC<IProps> = ({ url, scope, data }) => {
+const SearchResultRelatedList: React.FC<IProps> = ({
+  url,
+  scope,
+  data,
+  title,
+}) => {
   const recordLinks = (orderedItems: Array<IOrderedItems>): any =>
     orderedItems.map((item, ind: number) => {
       const { id } = item
@@ -70,7 +80,11 @@ const SearchResultRelatedList: React.FC<IProps> = ({ url, scope, data }) => {
               }`,
             }}
             onClick={() =>
-              pushSiteImproveEvent('Search Results Link', 'Selected', linkLabel)
+              pushSiteImproveEvent(
+                'Search Link',
+                'Selected',
+                `Accordion ${title}`,
+              )
             }
             data-testid="search-related-list-link"
           >