+++ client/src/test/unit/features/common/RelatedListSearchLink.spec.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react'
 import React from 'react'
 import { BrowserRouter } from 'react-router-dom'
 
-import SemanticSearchLink from '../../../../features/common/SemanticSearchLink'
+import RelatedListSearchLink from '../../../../features/relatedLists/RelatedListSearchLink'
 
 const mockId = 'test'
 const mockScope = 'item'
@@ -17,11 +17,11 @@ const mockCriteria = {
   ],
 }
 
-describe('SemanticSearchLink', () => {
+describe('RelatedListSearchLink', () => {
   it('renders', async () => {
     render(
       <BrowserRouter>
-        <SemanticSearchLink
+        <RelatedListSearchLink
           scope={mockScope}
           criteria={mockCriteria}
           id={mockId}
@@ -31,14 +31,14 @@ describe('SemanticSearchLink', () => {
       </BrowserRouter>,
     )
 
-    const link = screen.getByTestId(`semantic-search-link-${mockId}`)
+    const link = screen.getByTestId(`related-list-search-link-${mockId}`)
     expect(link).toBeInTheDocument()
   })
 
   it('renders with correct href', async () => {
     render(
       <BrowserRouter>
-        <SemanticSearchLink
+        <RelatedListSearchLink
           scope={mockScope}
           criteria={mockCriteria}
           id={mockId}
@@ -48,7 +48,7 @@ describe('SemanticSearchLink', () => {
       </BrowserRouter>,
     )
 
-    const link = screen.getByTestId(`semantic-search-link-${mockId}`)
+    const link = screen.getByTestId(`related-list-search-link-${mockId}`)
     expect(link).toHaveAttribute(
       'href',
       '/view/results/objects?q={"AND":[{"createdBy":{"id":"test"}}]}&openSearch=false',
@@ -58,7 +58,7 @@ describe('SemanticSearchLink', () => {
   it('renders the text if total and label are provided', async () => {
     render(
       <BrowserRouter>
-        <SemanticSearchLink
+        <RelatedListSearchLink
           scope={mockScope}
           criteria={mockCriteria}
           id={mockId}
@@ -69,14 +69,14 @@ describe('SemanticSearchLink', () => {
       </BrowserRouter>,
     )
 
-    const link = screen.getByTestId(`semantic-search-link-${mockId}`)
+    const link = screen.getByTestId(`related-list-search-link-${mockId}`)
     expect(link).toHaveTextContent('Show all 10 item results')
   })
 
   it('renders with correct text if totat and label are not provided', async () => {
     render(
       <BrowserRouter>
-        <SemanticSearchLink
+        <RelatedListSearchLink
           scope={mockScope}
           criteria={mockCriteria}
           id={mockId}
@@ -85,7 +85,7 @@ describe('SemanticSearchLink', () => {
       </BrowserRouter>,
     )
 
-    const link = screen.getByTestId(`semantic-search-link-${mockId}`)
+    const link = screen.getByTestId(`related-list-search-link-${mockId}`)
     expect(link).toHaveTextContent('Show all results')
   })
 })